refactor(email_service/send_email): split replace_placeholders function to template_util.py (SCRUM-73)
